[Decomp][Spell] Cohesion split: extract global cooldown to SpellCooldown.cpp
