+++ src/game/WorldHandlers/Spell.cpp
@@ -2982,101 +2982,6 @@ void Spell::ClearCastItem()
 }
 
 /**
- * @brief Checks whether the spell is currently blocked by the global cooldown.
- *
- * @return True if global cooldown is active; otherwise, false.
- */
-bool Spell::HasGlobalCooldown()
-{
-    // global cooldown have only player or controlled units
-    if (m_caster->GetCharmInfo())
-    {
-        return m_caster->GetCharmInfo()->GetGlobalCooldownMgr().HasGlobalCooldown(m_spellInfo);
-    }
-    else if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        return ((Player*)m_caster)->GetGlobalCooldownMgr().HasGlobalCooldown(m_spellInfo);
-    }
-    else
-    {
-        return false;
-    }
-}
-
-/**
- * @brief Starts the global cooldown for the caster when applicable.
- */
-void Spell::TriggerGlobalCooldown()
-{
-    int32 gcd = m_spellInfo->GetStartRecoveryTime();
-    if (!gcd)
-    {
-        return;
-    }
-
-    // global cooldown can't leave range 1..1.5 secs (if it it)
-    // exist some spells (mostly not player directly casted) that have < 1 sec and > 1.5 sec global cooldowns
-    // but its as test show not affected any spell mods.
-    if (gcd >= 1000 && gcd <= 1500)
-    {
-        // gcd modifier auras applied only to self spells and only player have mods for this
-        if (m_caster->GetTypeId() == TYPEID_PLAYER)
-        {
-            ((Player*)m_caster)->ApplySpellMod(m_spellInfo->Id, SPELLMOD_GLOBAL_COOLDOWN, gcd, this);
-        }
-
-        // apply haste rating
-        gcd = int32(float(gcd) * m_caster->GetFloatValue(UNIT_MOD_CAST_SPEED));
-
-        if (gcd < 1000)
-        {
-            gcd = 1000;
-        }
-        else if (gcd > 1500)
-        {
-            gcd = 1500;
-        }
-    }
-
-    // global cooldown have only player or controlled units
-    if (m_caster->GetCharmInfo())
-    {
-        m_caster->GetCharmInfo()->GetGlobalCooldownMgr().AddGlobalCooldown(m_spellInfo, gcd);
-    }
-    else if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)m_caster)->GetGlobalCooldownMgr().AddGlobalCooldown(m_spellInfo, gcd);
-    }
-}
-
-/**
- * @brief Cancels the global cooldown started by the current generic spell cast.
- */
-void Spell::CancelGlobalCooldown()
-{
-    if (!m_spellInfo->GetStartRecoveryTime())
-    {
-        return;
-    }
-
-    // cancel global cooldown when interrupting current cast
-    if (m_caster->GetCurrentSpell(CURRENT_GENERIC_SPELL) != this)
-    {
-        return;
-    }
-
-    // global cooldown have only player or controlled units
-    if (m_caster->GetCharmInfo())
-    {
-        m_caster->GetCharmInfo()->GetGlobalCooldownMgr().CancelGlobalCooldown(m_spellInfo);
-    }
-    else if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)m_caster)->GetGlobalCooldownMgr().CancelGlobalCooldown(m_spellInfo);
-    }
-}
-
-/**
  * @brief Resolves effective radius, chain target count, and target cap modifiers for an effect.
  *
  * @param spellEffect The spell effect entry being evaluated.
